requirements.txt
astroid==2.3.1
Babel==2.7.0
colorama==0.4.1
Django==2.2.6
django-ckeditor==5.7.1
django-hitcount==1.3.1
django-js-asset==1.2.2
django-mathfilters==0.4.0
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
django-scheduler==0.9.0
djangorestframework==3.10.3
icalendar==4.0.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
phonenumbers==8.10.20
Pillow==6.2.0
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
